Refactored slightly to implement dockerized container.
