+++ src/interfaces/cleverSections.ts
@@ -1,4 +1,7 @@
-import { CleverGradeType, CleverSubjectType } from '../../deps.ts';
+import {
+  CleverGradeType,
+  CleverSubjectType,
+} from '../lib/clever_library/mod.ts';
 import { Grades } from './enumGrades.ts';
 import { Subjects } from './enumSubjects.ts';
 import { IRumbleWithSectionInfo } from './rumbles.ts';
