+++ src/loaders/postgres.ts
@@ -4,15 +4,19 @@ import env from '../config/env.ts';
 export default async () => {
   console.log('Connecting to DB...');
 
-  const db = await connect({
-    type: 'postgres',
-    ...env.DB_CONFIG,
-  });
+  try {
+    const db = await connect({
+      type: 'postgres',
+      ...env.DB_CONFIG,
+    });
 
-  console.log('Testing DB connection...');
-  await db.query('SELECT * FROM users');
+    console.log('Testing DB connection...');
+    await db.query('SELECT * FROM users');
 
-  console.log('DB connected!');
+    console.log('DB connected!');
 
-  return db;
+    return db;
+  } catch (err) {
+    console.log(err.message, err);
+  }
 };
